pyproject.toml
[tool.poetry]
name = "bewise"
version = "0.1.0"
description = ""
authors = ["ilderebezov <ilderebezov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
SQLAlchemy = "^1.4.36"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.3"
Flask = "^2.1.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
